Add support for updating records
This adds support for updating records in place, using new `update` methods added to the record and record manager base classes.

The `update` method has the same smart evaluation of field names based on record classes that the `create` method has. This allows ID or object field names to be used for model refs, and field aliases.