Use lazy evaluation
Refactor the record manager `list` and `search` methods, and other methods on record managers, so that it is possible to use lazy evaluation.

The `list` and `search` methods now return `ListRecords` and `SearchRecords` objects, which do not perform any requests or processing at first. Calling the `as_ids`, `as_dicts` or `as_records` methods on these objects invokes the API request, but instead of returning lists as with the previous behaviour, generators are returned instead so the user can choose between performing per-item processing lazily on each iteration, or converting to a list to eagerly evaluate everything.

This removes the `as_id` and `as_dict` optional parameters from the `list` and `search` methods, which were not possible to express correctly with generators due to bugs in how Mypy handles conditional overloads. Returning a common object that contains separate methods with different, static return types is easier to make compliant with static type checkers.