Add support for managing attachments (#11)
Add a new manager for attachment records (`ir.attachment`) in Odoo.

These are intended to be used for uploading/downloading attachments to/from invoices.
Invoice attachments will then be attached to invoice emails sent out to customers.

The contents of the attachments won't be fetched when querying them from Odoo by default;
instead it is intended that the separate `download` method be used to download the attachment
contents separately. An `upload` method is also available, to provide an easier to use interface
for uploading attachments.

Co-authored-by: Callum Dickinson <callum.dickinson@catalystcloud.nz>